refactor: convert tab system to real Next.js pages
Replace the [[...tab]] catch-all with individual pages under a
(dashboard) route group. Each tab is now a separate route with
its own code-split JS bundle. Shared UI (intro, tab nav, global
filters) lives in the dashboard layout.

- Add tab-nav.tsx: Link-based navigation replacing Radix Tabs
- Add dashboard-shell.tsx: client wrapper for shared providers
- Add per-tab pages with co-located metadata
- Delete chart-tabs.tsx, page-content.tsx, [[...tab]]/page.tsx
- Add tabMetadata() helper to tab-meta.ts
