+++ packages/app/src/app/(dashboard)/layout.tsx
@@ -1,0 +1,5 @@
+import { DashboardShell } from '@/components/dashboard-shell';
+
+export default function DashboardLayout({ children }: { children: React.ReactNode }) {
+  return <DashboardShell>{children}</DashboardShell>;
+}
